refactor(hotkey): split hotkey_config.rs into per-type submodules
Extract the 1466-line file into hotkey_config/ with one sub-module per
type: key_code.rs (KeyCode + Display + from_name), key_combo.rs (KeyCombo
+ impls + Display + FromStr), action.rs (HotkeyAction + ALL + RESERVED +
per-variant methods), config.rs (HotkeyConfig + binding map + lookup +
TOML), tests.rs (the existing tests, dedented). mod.rs re-exports each
public type so all crate::types::hotkey_config::Foo paths still compile.

HotkeyConfig.bindings drops from private to pub(super) so the tests
sub-module can still inspect the map (the original tests touched it
directly via the same-file rule).

This is a baseline split — adding a HotkeyAction variant still touches
the seven match arms inside action.rs (display_name, description,
category, default_binding, to_toml_key, from_toml_key) plus the ALL
constant. Compiler exhaustiveness catches five of those, but the ALL
list and from_toml_key catch-all are silent. A follow-up commit
consolidates them under a single declaration via macro.

No behavioral change. cargo build / clippy --all-targets -D warnings /
nightly fmt --check / test --workspace all clean (1021 tests pass).
